+++ src/RadialAreaChart/RadialAreaSeries/RadialArea.tsx
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react';
 import { ChartInternalShallowDataShape } from '../../common/data';
 import { radialArea, curveCardinalClosed, curveLinear } from 'd3-shape';
-import { RadialGradient, RadialGradientProps } from '../../common/gradients';
+import { RadialGradient, RadialGradientProps } from '../../common/Gradient';
 import { CloneElement } from '../../common/utils';
 import { PosedRadialArea } from './PosedRadialArea';
 import { RadialInterpolationTypes } from '../../common/utils/interpolation';