+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - [BREAKING] Update `meta` property on `ChartData` type to `metadata`
 - [BREAKING] Update `TooltipArea` namespace to just `Tooltip`
 - [BREAKING] Remove internal `getGroupDomain` and `getDeepGroupDomain` in favor of `uniqueBy`
+- [BREAKING] Rename internal exports `gradients` and `masks` to just `Gradient` and `Mask`
 - [Feature] Year Calendar Heatmaps
 - [Feature] Month Calendar Heatmaps
 - [Feature] Add ability to pass empty heatmap colors